@@ -2,8 +2,8 @@
 export const questionLoadDelay = 2000;
 export const breakTime = 300000;
 export const autoWrongGuess = 60000;
-export const correctColor = 'green';
-export const incorrectColor = 'red';
+export const correctColor = '#009900';
+export const incorrectColor = '#e50000';
 
 /**
  * Key that was used to encrypt the questions
@@ -17,3 +17,4 @@ export const resultsURL = 'results';  // URL to web api
 export const resultsCheckURL = 'getResult';
 export const dateNowURL = 'dateNow';
 export const dateStartURL = 'dateStart';
+export const canStartURL = 'canStart';
