@@ -4,25 +4,25 @@ import { Router } from '@angular/router';
 import { questionLoadDelay, autoWrongGuess, breakTime } from '../constants';
 import { Question } from '../question';
 import { QuestionService } from '../question.service';
-import { Result } from '../result';
-import { ResultService } from '../result.service';
+import { Team } from '../team';
+import { TeamService } from '../team.service';
 import { TimerComponent } from '../timer/timer.component';
 
 @Component({
-  selector: 'vq-game',
+  selector: 'app-game',
   templateUrl: 'game.component.html',
   styleUrls: ['game.component.css']
 })
 export class GameComponent implements OnInit, AfterViewInit {
 
-	@ViewChild(TimerComponent)
-	private timerComponent: TimerComponent;
+  @ViewChild(TimerComponent)
+  private timerComponent: TimerComponent;
 
   /**
    * Whether or not a breakStarted is currently in progress. True if in progress; false otherwise
    * @type {boolean}
    */
-  breakStarted: boolean = false;
+  breakStarted = false;
   /**
    * Used to store the setTimeout() variable so that we can later call clearTimeout()
    */
@@ -36,7 +36,7 @@ export class GameComponent implements OnInit, AfterViewInit {
    * or chose the correct answer; false otherwise
    * @type {boolean}
    */
-  finished: boolean = false;
+  finished = false;
   /**
    * Index of the current Question
    * @type {number}
@@ -46,55 +46,57 @@ export class GameComponent implements OnInit, AfterViewInit {
    * The user's current point score
    * @type {number}
    */
-  points: number = 0;
+  points = 0;
   /**
    * Whether or not this is a practice round; True if a practice round; false otherwise
    */
-  practice: boolean = false;
+  practice = false;
   /**
    * The Questions to display
    */
-	questions: Question[];
+  questions: Question[];
   /**
    * The Result object which contains identifying information about the user
    */
-  result: Result;
+  team: Team;
   /**
    * Whether or not the user is allowed a second guess on the current question. True when allowed; false otherwise
    * @type {boolean}
    */
-	secondTryAllowed: boolean = true;
-	pointsGained: number = 0;
+  secondTryAllowed = true;
+  pointsGained = 0;
 
-	constructor(private router: Router,
-		private questionService: QuestionService,
-    private resultService: ResultService) { }
+  constructor(private router: Router,
+              private questionService: QuestionService,
+              private teamService: TeamService) { }
 
   /**
    * Called when there are no more questions to serve, i.e. when the game is over.
    */
   gameOver() {
-    this.result.points = this.points;
-    this.result.timeEnded = Date.now();
-    this.resultService.setPractice(this.practice);
-    this.resultService.setResult(this.result);
+    this.team.points = this.points;
+    this.team.timeEnded = Date.now();
+    this.teamService.setPractice(this.practice);
+    this.teamService.setTeam(this.team);
     this.router.navigate(['/gameover']);
   }
   /**
    * Retrieves the practice questions through QuestionService
    */
   getPracticeQuestions() {
-    this.questionService.getPracticeQuestions()
-      .then(questions => this.questions = questions)
-      .then(() => this.currentQuestion = this.questions[this.index]);
+    this.questionService.getPracticeQuestions().subscribe(questions => {
+      this.questions = questions;
+      this.currentQuestion = this.questions[this.index];
+    });
   }
   /**
    * Retrieves the questions through QuestionService
    */
   getQuestions() {
-    this.questionService.getQuestions()
-      .then(questions => this.questions = questions)
-      .then(() => this.currentQuestion = this.questions[this.index]);
+    this.questionService.getQuestions().subscribe(questions => {
+        this.questions = questions;
+        this.currentQuestion = this.questions[this.index];
+      });
   }
   /**
    * Loads the next question if it exists, and if not, calls gameOver(). Also in charge of initiating breaks, which
@@ -103,50 +105,48 @@ export class GameComponent implements OnInit, AfterViewInit {
   loadQuestion() {
     // save result
     this.pointsGained = 0;
-    this.result.currentQuestion = this.index + 1;
-    this.result.points = this.points;
+    this.team.currentQuestion = this.index + 1;
+    this.team.points = this.points;
     if (!this.practice) {
-      this.resultService.save(this.result);
+      this.teamService.save(this.team).subscribe(() => console.log('team data saved'));
     }
     ++this.index;
     if (this.questions[this.index]) {
-      if (this.index % Math.floor(this.questions.length/3) === 0 && this.index != Math.floor(this.questions.length/3)*3) {
+      if (this.index % Math.floor(this.questions.length / 3) === 0 && this.index != Math.floor(this.questions.length / 3) * 3 ) {
         this.breakStarted = true;
         this.restartTimer();
         // Prevents on breakStarted menu from continuing to reset after 60 seconds. Essentially undoes onStarted()
         this.timerComponent.clearInterval();
-        this.breakEnd = setTimeout(()=> {
+        this.breakEnd = setTimeout(() => {
           this.breakStarted = false;
           this.questionHelper();
         }, breakTime);
-      }
-      else {
+      } else {
         this.questionHelper();
       }
-    }
-    else {
+    } else {
       this.gameOver();
     }
   }
   ngAfterViewInit() {
     // sets up the seconds() method to actually get the time from the TimerComponent
     // Dubious if this should actually be in AfterViewInit()
-    setTimeout(() => this.seconds = () => this.timerComponent.mseconds/1000, 0);
+    setTimeout(() => this.seconds = () => this.timerComponent.mseconds / 1000, 0);
     this.startTimer();
   }
   ngOnInit() {
     // Necessarily executed in order?
-    this.result = this.resultService.getResult();
-    if (!this.result.timeStarted) {
-      this.result.timeStarted = Date.now();
+    this.team = this.teamService.getTeam();
+    if (!this.team.timeStarted) {
+      this.team.timeStarted = Date.now();
     }
-    if (this.result.currentQuestion) {
-      this.index = this.result.currentQuestion;
+    if (this.team.currentQuestion) {
+      this.index = this.team.currentQuestion;
     }
-    if (this.result.points) {
-      this.points = this.result.points;
+    if (this.team.points) {
+      this.points = this.team.points;
     }
-    this.practice = this.result._id === 'practice';
+    this.practice = this.team._id === 'practice';
     if (!this.practice) {
       this.getQuestions();
     } else {
@@ -172,7 +172,7 @@ export class GameComponent implements OnInit, AfterViewInit {
         this.pointsGained = 2;
       }
       this.points += this.pointsGained;
-      setTimeout(()=>this.loadQuestion(), questionLoadDelay);
+      setTimeout(() => this.loadQuestion(), questionLoadDelay);
     } else {
       if (this.secondTryAllowed) {
         this.secondTryAllowed = false;
@@ -180,7 +180,7 @@ export class GameComponent implements OnInit, AfterViewInit {
       } else {
         this.finished = true;
         this.stopTimer();
-        setTimeout(()=>this.loadQuestion(), questionLoadDelay);
+        setTimeout(() => this.loadQuestion(), questionLoadDelay);
       }
     }
   }
@@ -198,44 +198,44 @@ export class GameComponent implements OnInit, AfterViewInit {
    * counted as an automatic wrong guess.
    */
   onStarted() {
-    this.timerComponent.setInterval(()=> {
+    this.timerComponent.setInterval(() => {
       if (this.secondTryAllowed) {
         this.secondTryAllowed = false;
         this.restartTimer();
       } else {
         this.finished = true;
         this.stopTimer();
-        setTimeout(()=>this.loadQuestion(), questionLoadDelay);
+        setTimeout(() => this.loadQuestion(), questionLoadDelay);
       }
     }, autoWrongGuess);
   }
   /**
    * Stops, resets, and then starts the timer
    */
-  restartTimer() { this.timerComponent.restartTimer() }
+  restartTimer() { this.timerComponent.restartTimer(); }
   /**
    * Resets the timer
    */
-  resetTimer() { this.timerComponent.resetTimer() }
+  resetTimer() { this.timerComponent.resetTimer(); }
   /**
    * The current number of seconds on the timer
    * @returns {number}
    *  Current number of seconds on timer
    */
-	seconds() { return 0; }
+  seconds() { return 0; }
   /**
    * Starts the timer
    */
-	startTimer() { this.timerComponent.startTimer() }
+  startTimer() { this.timerComponent.startTimer(); }
   /**
    * Stops the timer but does not reset it
    */
-	stopTimer() { this.timerComponent.stopTimer() }
+  stopTimer() { this.timerComponent.stopTimer(); }
 
   /**
    * A little helper that loads the next question.
    */
-	private questionHelper() {
+  private questionHelper() {
     this.currentQuestion = this.questions[this.index];
     this.restartTimer();
     this.secondTryAllowed = true;
