@@ -1,11 +1,11 @@
-import { Answer } from './answer';
-
 /**
  * Holds a single question
  */
+import { Answer } from './answer';
+
 export class Question {
   constructor(public question: string,
               public category: string,
-              public answers: Answer[],
+              public answers: Answer[]
   ) {}
 }
