@@ -1,8 +1,5 @@
-// Typings reference file, see links for more information
-// https://github.com/typings/typings
-// https://www.typescriptlang.org/docs/handbook/writing-declaration-files.html
-
-declare var System: any;
-declare var module: { id: string };
-declare var require: any;
-
+/* SystemJS module definition */
+declare var module: NodeModule;
+interface NodeModule {
+  id: string;
+}
